requirements.txt
setuptools>=69.0.0
wheel>=0.42.0
python-binance
google-generativeai
aiohttp
pyyaml
numpy>=1.24.3
asyncio
requests
python-telegram-bot
pytest
pandas>=2.0.3
matplotlib>=3.7.1
scikit-learn>=1.3.0
joblib>=1.2.0
scipy>=1.11.1